Completely remove regex and code cleanup
